fix: apply reviewer findings for Phase 5 LangGraph pipeline
- C1: multi-event idempotency scoped to (source_doc, business_name)
- C2: LLM call records sent once per page, not per extracted event
- W1: add extraction stats to CLI summary printout
- W2: set final_status=partial when extraction_errors > 0
- W3: 404 guard for unknown source_document_id in save_candidate_event
- W4: remove source_document_id from CandidateEventCreate body schema
- W5: remove DATABASE_URL from worker service in docker-compose.yml
- W6: add TestSaveCandidateEvent unit tests (404, irrelevant, dup, happy path)
- N1: use Decimal for cost_usd in LLMCall ORM model
- N2: fix build_extraction_graph return type to CompiledStateGraph
- N3: schema_class: type[BaseModel] in _call_openai
- N4: remove unused monkeypatch params from graph tests
- N5: remove unused LLMCallCreate import from integration tests
- N6: add 404 test for unknown source_document_id in integration tests
- N7: log debug message for unknown model in _estimate_cost
- Update README and 09_phases.md to mark Phase 5 complete
- 198 tests passing

+++ backend/app/models/llm_calls.py
@@ -8,6 +8,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sqlalchemy import DateTime, Numeric, String, Text
@@ -36,7 +37,7 @@ class LLMCall(Base):
     prompt_tokens: Mapped[int | None] = mapped_column(nullable=True)
     completion_tokens: Mapped[int | None] = mapped_column(nullable=True)
     total_tokens: Mapped[int | None] = mapped_column(nullable=True)
-    cost_usd: Mapped[float | None] = mapped_column(Numeric(10, 6), nullable=True)
+    cost_usd: Mapped[Decimal | None] = mapped_column(Numeric(10, 6), nullable=True)
     latency_ms: Mapped[int | None] = mapped_column(nullable=True)
     status: Mapped[str] = mapped_column(
         String(20), nullable=False, default="success"
